refactor(task): 将SequenceClassificationTask的运行逻辑迁移到Task类中
